Updated project structure + Fixed macOS/watchOS/tvOS issues + Bumped version to 0.0.3
