+++ QonversionSandwich.podspec
@@ -1,6 +1,7 @@
 Pod::Spec.new do |s|
+  excluded_files = ['ios/sandwich/AutomationsSandwich.swift', 'ios/sandwich/AutomationsEventListener.swift', 'ios/sandwich/AutomationsMappers.swift']
   s.name         = 'QonversionSandwich'
-  s.version      = '0.0.2'
+  s.version      = '0.0.3'
   s.summary      = 'qonversion.io'
   s.swift_version = '5.0'
   s.description  = <<-DESC
@@ -12,6 +13,10 @@ Pod::Spec.new do |s|
   s.source       = { :git => 'https://github.com/qonversion/sandwich-sdk.git', :tag => s.version.to_s }
   s.framework    = 'StoreKit'
   s.platforms    = { :ios => "9.0", :osx => "10.12", :tvos => "9.0", :watchos => "6.2" }
+  
+  s.osx.exclude_files         = excluded_files
+  s.tvos.exclude_files        = excluded_files
+  s.watchos.exclude_files     = excluded_files
 
   s.source_files = 'ios/sandwich/**/*.{h,m,swift}'
   s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
